+++ docker-compose-manager-back/docker-compose-manager-back/Migrations/20260214223639_UpdateAdminEmailToNull.cs
@@ -1,0 +1,32 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace docker_compose_manager_back.Migrations
+{
+    /// <inheritdoc />
+    public partial class UpdateAdminEmailToNull : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.UpdateData(
+                table: "Users",
+                keyColumn: "Id",
+                keyValue: 1,
+                columns: new[] { "Email", "MustAddEmail" },
+                values: new object[] { null, true });
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.UpdateData(
+                table: "Users",
+                keyColumn: "Id",
+                keyValue: 1,
+                columns: new[] { "Email", "MustAddEmail" },
+                values: new object[] { "admin@localhost", false });
+        }
+    }
+}